Add files via upload
This was our first programming homework for this course, and I really enjoyed figuring out how to optomize my solutions for answers 1 and 2 into 1 line of code. 

After quite a bit of trial and error I got problem 5 to work, I was having an issue where by removing a non-prime from my list, it caused the indexing in my for loop to actually skip the next element (As it was in the previously removed element position now), and learned about slicing and copying a list in Python to reach a solution.